Refactor to remove errors
Also lowered error level to get rid of -Wpadding and -Wunsuffixed-...
