+++ src/lib/common_headers/include/unused/mark_unused.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#define UNUSED(x) ((void)((x)))
