+++ src/mrc/mimir/meson.build
@@ -3,6 +3,15 @@ mimir_include = include_directories('include',
     '../../lib/histogram/include',
 )
 
+mimir_buckets_lib = library('mimir_buckets_lib',
+    'buckets.c',
+    include_directories : [
+        common_headers_include,
+        mimir_include,
+    ],
+    dependencies : glib_dep,
+)
+
 mimir_lib = library('mimir_lib',
     'mimir.c',
     include_directories : [
@@ -15,4 +24,5 @@ mimir_lib = library('mimir_lib',
         basic_tree_lib,
         fractional_histogram_lib,
         sleator_tree_lib,
+        mimir_buckets_lib,
     ])
