+++ test/unit_test/lib_test/histogram_test/histogram_test.c
@@ -6,31 +6,33 @@
 #include "histogram/basic_histogram.h"
 
 #include "test/mytester.h"
+#include "unused/mark_unused.h"
 
 // NOTE These are 100 random integers in the range 0..=10. Generated with Python
 //      script:
 //      import random; x = [random.randint(0, 10) for _ in range(100)]; print(x)
 const uint64_t random_values_0_to_11[100] = {
-    2, 3, 2,  5, 0,  1, 7, 9, 4, 2, 10, 3,  1, 10, 10, 5, 10, 6, 5, 0, 6, 4, 2, 9, 7,
-    2, 2, 5,  3, 9,  6, 0, 1, 1, 6, 1,  6,  7, 5,  0,  0, 10, 8, 3, 1, 2, 6, 7, 3, 10,
-    8, 6, 10, 6, 6,  2, 6, 0, 7, 9, 6,  10, 1, 10, 2,  6, 2,  7, 8, 8, 6, 0, 7, 3, 1,
-    1, 2, 10, 3, 10, 5, 5, 0, 7, 9, 8,  0,  7, 6,  9,  4, 9,  4, 8, 3, 6, 5, 3, 2, 9};
+    2, 3,  2, 5,  0, 1, 7, 9, 4, 2,  10, 3, 1,  10, 10, 5, 10, 6,  5, 0,
+    6, 4,  2, 9,  7, 2, 2, 5, 3, 9,  6,  0, 1,  1,  6,  1, 6,  7,  5, 0,
+    0, 10, 8, 3,  1, 2, 6, 7, 3, 10, 8,  6, 10, 6,  6,  2, 6,  0,  7, 9,
+    6, 10, 1, 10, 2, 6, 2, 7, 8, 8,  6,  0, 7,  3,  1,  1, 2,  10, 3, 10,
+    5, 5,  0, 7,  9, 8, 0, 7, 6, 9,  4,  9, 4,  8,  3,  6, 5,  3,  2, 9};
 
-#define ASSERT_TRUE_OR_RETURN_FALSE(r, msg, histogram_ptr)                                         \
-    do {                                                                                           \
-        if (!(r)) {                                                                                \
-            /* Clean up resources */                                                               \
-            basic_histogram__destroy((histogram_ptr));                                             \
-            printf("[ERROR] %s:%d %s\n", __FILE__, __LINE__, (msg));                               \
-            /* NOTE This assertion is for debugging purposes so that we have a                     \
-            finer grain understanding of where the failure occurred. */                            \
-            assert((r) && (msg));                                                                  \
-            return false;                                                                          \
-        }                                                                                          \
+#define ASSERT_TRUE_OR_RETURN_FALSE(r, msg, histogram_ptr)                     \
+    do {                                                                       \
+        if (!(r)) {                                                            \
+            /* Clean up resources */                                           \
+            basic_histogram__destroy((histogram_ptr));                         \
+            printf("[ERROR] %s:%d %s\n", __FILE__, __LINE__, (msg));           \
+            /* NOTE This assertion is for debugging purposes so that we have a \
+            finer grain understanding of where the failure occurred. */        \
+            assert((r) && (msg));                                              \
+            return false;                                                      \
+        }                                                                      \
     } while (0)
 
-bool
-test_basic_histogram()
+static bool
+test_basic_histogram(void)
 {
     struct BasicHistogram hist;
 
@@ -53,7 +55,8 @@ test_basic_histogram()
     }
     // NOTE Histogram oracle generated by getting the random values generated
     //      above and running:
-    //      y = [x.count(i) for i in range(10)]; print(y); z = x.count(10); print(z)
+    //      y = [x.count(i) for i in range(10)]; print(y); z = x.count(10);
+    //      print(z)
     const uint64_t histogram_oracle[] = {9, 9, 12, 9, 4, 8, 15, 9, 6, 8};
     const uint64_t length_oracle = 10;
     const uint64_t false_infinity_oracle = 11;
@@ -65,7 +68,9 @@ test_basic_histogram()
                                     "histogram should match oracle",
                                     &hist);
     }
-    ASSERT_TRUE_OR_RETURN_FALSE(hist.length == length_oracle, "length should match oracle", &hist);
+    ASSERT_TRUE_OR_RETURN_FALSE(hist.length == length_oracle,
+                                "length should match oracle",
+                                &hist);
     ASSERT_TRUE_OR_RETURN_FALSE(hist.false_infinity == false_infinity_oracle,
                                 "false_infinity should match oracle",
                                 &hist);
@@ -79,8 +84,10 @@ test_basic_histogram()
 }
 
 int
-main(int argc, char *argv[])
+main(int argc, char **argv)
 {
+    UNUSED(argc);
+    UNUSED(argv);
     ASSERT_FUNCTION_RETURNS_TRUE(test_basic_histogram());
     return 0;
 }
