+++ test/unit_test/lib_test/histogram_test/meson.build
@@ -1,6 +1,6 @@
 histogram_test_exe = executable('histogram_test_exe',
     'histogram_test.c',
-    include_directories : [mytester_include, histogram_include],
+    include_directories : [common_headers_include, mytester_include, histogram_include],
     dependencies : glib_dep,
     link_with : [basic_histogram_lib])
