+++ test/unit_test/lib_test/priority_queue_test/priority_queue_test.c
@@ -23,13 +23,17 @@ test_splay_priority_queue(void)
     for (Hash64BitType i = 0; i < 10; ++i) {
         // The pq should not be full until we reach the end of this loop!
         g_assert_false(splay_priority_queue__is_full(&pq));
-        g_assert_true(splay_priority_queue__insert_if_room(&pq, (Hash64BitType)i, (EntryType)i));
+        g_assert_true(splay_priority_queue__insert_if_room(&pq,
+                                                           (Hash64BitType)i,
+                                                           (EntryType)i));
     }
     // The pq should now be full!
     g_assert_true(splay_priority_queue__is_full(&pq));
     // This is kind of a hack but it's because I am making my own macro library
     // and don't want to have to define a TRUE and FALSE version.
-    g_assert_false(splay_priority_queue__insert_if_room(&pq, (Hash64BitType)10, (EntryType)10));
+    g_assert_false(splay_priority_queue__insert_if_room(&pq,
+                                                        (Hash64BitType)10,
+                                                        (EntryType)10));
 
     // Get maximum
     Hash64BitType hash = splay_priority_queue__get_max_hash(&pq);
@@ -55,7 +59,8 @@ test_splay_priority_queue(void)
     g_assert_true(splay_priority_queue__remove(&pq, hash, &entry_1));
     hash = splay_priority_queue__get_max_hash(&pq);
     g_assert_cmpuint(hash, ==, 7);
-    g_assert_true((entry_0 == 8 && entry_1 == 9) || (entry_0 == 9 && entry_1 == 8));
+    g_assert_true((entry_0 == 8 && entry_1 == 9) ||
+                  (entry_0 == 9 && entry_1 == 8));
 
     for (uint64_t i = 0; i < 7; ++i) {
         Hash64BitType expected_max_hash = 7 - i;
@@ -63,23 +68,23 @@ test_splay_priority_queue(void)
         hash = splay_priority_queue__get_max_hash(&pq);
         g_assert_cmpuint(hash, ==, expected_max_hash);
         // Remove
-        g_assert_true(splay_priority_queue__remove(&pq, expected_max_hash, &entry));
+        g_assert_true(
+            splay_priority_queue__remove(&pq, expected_max_hash, &entry));
         g_assert_cmpuint(entry, ==, (EntryType)expected_max_hash);
         // Try another removal that should fail
         hash = splay_priority_queue__get_max_hash(&pq);
         g_assert_cmpuint(hash, ==, expected_max_hash - 1);
-        g_assert_false(splay_priority_queue__remove(&pq, expected_max_hash, &entry));
+        g_assert_false(
+            splay_priority_queue__remove(&pq, expected_max_hash, &entry));
     }
-
     splay_priority_queue__destroy(&pq);
 }
 
 int
-main(int argc, char *argv[])
+main(int argc, char **argv)
 {
     g_test_init(&argc, &argv, NULL);
-
-    g_test_add_func("/priority_queue_test/splay_priority_queue__test", test_splay_priority_queue);
-
+    g_test_add_func("/priority_queue_test/splay_priority_queue__test",
+                    test_splay_priority_queue);
     return g_test_run();
 }
